fastapi_dummy/requirements.txt
fastapi ~= 0.81
robotframework ~= 5.0
requests ~= 2.28